[test] refactor test_manuf_debug_unlock test
This refactors the manuf debug unlock FPGA test into its own file.
Additionally, it updates the test to make use of the JTAG mailbox
utility functions.
